+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - removed migration framework. we are adding migration feature to fastn core itself, so no need to
   maintain it here.
+- support for optional query parameters in `ft_sdk::Query`.
 
 ## 3rd June 2024